Handle Promise rejection properly
The previous version of this function returned a Promise that was not
awaited in a try block, so when the Promise rejected, the try block did
not catch it, resulting in a ENOENT when the error should have been
recovered.

I've used the .catch method of the Promise instead as I think it's more
readable than using the try/catch construct in this case.

Resolves #4